+++ src/globalEnum.ts
@@ -22,7 +22,7 @@ export type ReplaceAll<Str extends string, Sub extends string, NSub extends stri
 
 export const enum EDetail {
     inComment = 3,
-    inSkipSign2 = 4,
+    inSkipSign2 = 4, // old varName = v
     deepAdd = 5,
     deepSubtract = 6,
     hasDoubleSemicolon = 8,